okto-juvenile-bnb/requirements.txt
ccxt
web3
python-dotenv
keyring
pandas
termcolor
pyyaml
pyyaml
